help: Make value->html have a more comprehensive & intuitive output.
This adds reasonable printing for property and associative lists,
hash-tables, classes and sequences. It's recursive: all the
sub-objects will be recursively listed as exhaustively as the original
one.

The cost for this is losing references to list elements that used to
be there. This is not much of a loss, as we should have a better
object reference system, with an arbitrary number of objects being
introspectable and linkable to. Until then, we'll lack links to list
elements in value->html, but its output will be much more useful than
the previous one, even without links.